🐛 Wrong sd used in example
More generally, the residual variance used in first Alien example (5) differs to the default in later slides (25) (e.g. `simulate_Aliens()` func). Not major, but maybe we want to change one or the other so it is consistent. Intercept and effect of 'humans eaten' remains the same between these examples.